requirements.txt
certifi==2018.8.24
chardet==3.0.4
idna==2.7
pygeocoder==1.2.5
PySocks==1.6.8
requests==2.19.1
urllib3==1.23
